requirements.txt
django~=5.0.3
django_restframework
pyrebase
djangorestframework~=3.15.1